Only prepend root directory when pattern is not absolute
os.path.join conveniently will ignore preceding path parts if the later
path part is absolute, but instead of relying on that logic (in case it
ever changes, which hopefully it wouldn't), we should just check whether
we actually need to call os.path.join. We already have the info tracked,
it is just a matter of checking it.