Remove unnecessary directory check from glob
When under the `dir_only` conditional block, it can be assumed
that all directories returned are directories. If this assumption
is not true, it means there is a bug in the code. Eliminate the
redundant check of whether the file in question is a directory
in the main, non-magic, directory only loop.