Drop requirement for backrefs as the Unicode tables were never used
While we had intended to use the Unicode tables, we never actually did,
and no one ever noticed, not even me (until now). As we've stuck with
the ASCII tables all this time, and it has been more than sufficient,
port that logic into wcmatch directly and drop backrefs. Update docs
so there is no confusion.