Refactor non public parts of wcmatch
Rummage, which wcmatch was originally written for, overrides path
pattern compilation in a messy way. And that area of the code in
wcmatch is kind of messy as well. Clean this up to make this more
clear and concise. Prove backwards compatibility for Rummage so it
can transition off the old ugly way without breaking existing releases.

This does not affect the public API in any way, and any one mucking
with undocumented things is responsible for handling such changes.
Rummage is the only exception I take great care with as it is what
I originally wrote this project for.