Elaborate a little more on capturing groups
Explain the purpose of generating and returning regex patterns and
emphasize that even though extended glob patterns do return capturing
groups, that can be utilized, the main goal of the regex patterns is
for matching a file path generally. The focus is not to provide advanced
capturing withing these regex patterns, at least not currently.